Update Ch 07 - Regex Strip.py
Hi, 

I may be wrong but the function does not behave like the strip method of str type. In the strip method if an argument is provided the string it is called on is returned with the argument stripped only from the start or end of the string. The function above seems to work when no second argument is provided. When a second argument is provided it instead strips that argument from the string regardless of where in the string that argument is located. I may be wrong?

I adjusted line 25. I may be incorrect so please let me know if I am. I   have been learning a lot from the solutions you've posted to the chapter projects of this book. Thanks very much.